+++ Tunny/Solver/Algorithm.cs
@@ -20,16 +20,15 @@ using Tunny.Settings;
 using Tunny.Storage;
 using Tunny.Type;
 using Tunny.UI;
-using Tunny.Util;
 
 namespace Tunny.Solver
 {
     public class Algorithm
     {
-        public double[] XOpt { get; private set; }
+        public Parameter[] OptimalParameters { get; private set; }
         public EndState EndState { get; private set; }
 
-        private List<Variable> Variables { get; }
+        private List<VariableBase> Variables { get; }
         private bool HasConstraints { get; }
         private Objective Objective { get; }
         private TunnySettings Settings { get; }
@@ -37,7 +36,7 @@ namespace Tunny.Solver
         private Dictionary<string, FishEgg> FishEgg { get; }
 
         public Algorithm(
-            List<Variable> variables, bool hasConstraint, Objective objective, Dictionary<string, FishEgg> fishEgg,
+            List<VariableBase> variables, bool hasConstraint, Objective objective, Dictionary<string, FishEgg> fishEgg,
             TunnySettings settings,
             Func<ProgressState, int, TrialGrasshopperItems> evalFunc)
         {
@@ -63,14 +62,13 @@ namespace Tunny.Solver
             using (Py.GIL())
             {
                 Log.Debug("Wake Python GIL.");
-                PythonEngine.Exec("a == 1");
 
                 dynamic optuna = Py.Import("optuna");
                 dynamic sampler = SetSamplerSettings(samplerType, optuna, HasConstraints);
                 dynamic storage = Settings.Storage.CreateNewOptunaStorage(false);
                 dynamic artifactBackend = Settings.Storage.CreateNewOptunaArtifactBackend(false);
 
-                double[] parameter = null;
+                Parameter[] parameter = null;
                 TrialGrasshopperItems result = null;
 
                 if (CheckExistStudyParameter(Objective.Length, optuna, storage))
@@ -107,7 +105,7 @@ namespace Tunny.Solver
             Log.Debug("Initialize PythonEngine.");
         }
 
-        private void PreferentialOptimization(int nBatch, dynamic storage, dynamic artifactBackend, out double[] parameter, out TrialGrasshopperItems result, out dynamic study)
+        private void PreferentialOptimization(int nBatch, dynamic storage, dynamic artifactBackend, out Parameter[] parameter, out TrialGrasshopperItems result, out dynamic study)
         {
             if (Settings.Storage.Type != StorageType.Sqlite)
             {
@@ -128,7 +126,7 @@ namespace Tunny.Solver
             RunPreferentialOptimize(optInfo, nBatch, out parameter, out result);
         }
 
-        private void NormalOptimization(int nTrials, double timeout, string[] directions, dynamic sampler, dynamic storage, dynamic artifactBackend, out double[] parameter, out TrialGrasshopperItems result, out dynamic study)
+        private void NormalOptimization(int nTrials, double timeout, string[] directions, dynamic sampler, dynamic storage, dynamic artifactBackend, out Parameter[] parameter, out TrialGrasshopperItems result, out dynamic study)
         {
             study = CreateStudy(directions, sampler, storage);
             string[] objNickName = Objective.GetNickNames();
@@ -137,7 +135,7 @@ namespace Tunny.Solver
             RunOptimize(optInfo, out parameter, out result);
         }
 
-        private void HumanSliderInputOptimization(int nBatch, double timeout, string[] directions, dynamic sampler, dynamic storage, dynamic artifactBackend, out double[] parameter, out TrialGrasshopperItems result, out dynamic study)
+        private void HumanSliderInputOptimization(int nBatch, double timeout, string[] directions, dynamic sampler, dynamic storage, dynamic artifactBackend, out Parameter[] parameter, out TrialGrasshopperItems result, out dynamic study)
         {
             study = CreateStudy(directions, sampler, storage);
             string[] objNickName = Objective.GetNickNames();
@@ -220,13 +218,14 @@ namespace Tunny.Solver
             }
         }
 
-        private void SetResultValues(int nObjective, dynamic study, double[] parameter)
+        private void SetResultValues(int nObjective, dynamic study, Parameter[] parameter)
         {
             if (nObjective == 1 && Objective.HumanInTheLoopType != HumanInTheLoopType.Preferential)
             {
-                double[] values = (double[])study.best_params.values();
+                PyObject[] pyBestParams = study.best_params.values();
+                string[] values = pyBestParams.Select(x => x.ToString()).ToArray();
                 string[] keys = (string[])study.best_params.keys();
-                double[] opt = new double[Variables.Count];
+                var opt = new Parameter[Variables.Count];
 
                 for (int i = 0; i < Variables.Count; i++)
                 {
@@ -234,21 +233,21 @@ namespace Tunny.Solver
                     {
                         if (keys[j] == Variables[i].NickName)
                         {
-                            opt[i] = values[j];
+                            opt[i] = double.TryParse(values[j], out double num) ? new Parameter(num) : new Parameter(values[j]);
                         }
                     }
                 }
-                XOpt = opt;
+                OptimalParameters = opt;
             }
             else
             {
-                XOpt = parameter;
+                OptimalParameters = parameter;
             }
         }
 
-        private void RunOptimize(OptimizationHandlingInfo optInfo, out double[] parameter, out TrialGrasshopperItems result)
+        private void RunOptimize(OptimizationHandlingInfo optInfo, out Parameter[] parameter, out TrialGrasshopperItems result)
         {
-            parameter = new double[Variables.Count];
+            parameter = new Parameter[Variables.Count];
             result = new TrialGrasshopperItems();
             int trialNum = 0;
             DateTime startTime = DateTime.Now;
@@ -267,9 +266,9 @@ namespace Tunny.Solver
             SaveInMemoryStudy(optInfo.Storage);
         }
 
-        private void RunPreferentialOptimize(OptimizationHandlingInfo optInfo, int nBatch, out double[] parameter, out TrialGrasshopperItems result)
+        private void RunPreferentialOptimize(OptimizationHandlingInfo optInfo, int nBatch, out Parameter[] parameter, out TrialGrasshopperItems result)
         {
-            parameter = new double[Variables.Count];
+            parameter = new Parameter[Variables.Count];
             result = new TrialGrasshopperItems();
             int trialNum = 0;
             DateTime startTime = DateTime.Now;
@@ -292,9 +291,9 @@ namespace Tunny.Solver
             SaveInMemoryStudy(optInfo.Storage);
         }
 
-        private void RunHumanSidlerInputOptimize(OptimizationHandlingInfo optInfo, int nBatch, out double[] parameter, out TrialGrasshopperItems result)
+        private void RunHumanSidlerInputOptimize(OptimizationHandlingInfo optInfo, int nBatch, out Parameter[] parameter, out TrialGrasshopperItems result)
         {
-            parameter = new double[Variables.Count];
+            parameter = new Parameter[Variables.Count];
             result = new TrialGrasshopperItems();
             int trialNum = 0;
             DateTime startTime = DateTime.Now;
@@ -317,7 +316,7 @@ namespace Tunny.Solver
             SaveInMemoryStudy(optInfo.Storage);
         }
 
-        private TrialGrasshopperItems RunSingleOptimizeStep(OptimizationHandlingInfo optInfo, double[] parameter, int trialNum, DateTime startTime)
+        private TrialGrasshopperItems RunSingleOptimizeStep(OptimizationHandlingInfo optInfo, Parameter[] parameter, int trialNum, DateTime startTime)
         {
             dynamic trial;
             int progress = trialNum * 100 / optInfo.NTrials;
@@ -333,13 +332,7 @@ namespace Tunny.Solver
                 }
 
                 trial = optInfo.Study.ask();
-
-                for (int j = 0; j < Variables.Count; j++)
-                {
-                    parameter[j] = Variables[j].IsInteger
-                    ? trial.suggest_int(Variables[j].NickName, Variables[j].LowerBond, Variables[j].UpperBond, step: Variables[j].Epsilon)
-                    : trial.suggest_float(Variables[j].NickName, Variables[j].LowerBond, Variables[j].UpperBond, step: Variables[j].Epsilon);
-                }
+                SetOptimizationParameter(parameter, trial);
 
                 ProgressState pState = SetProgressState(optInfo, parameter, trialNum, startTime);
                 result = EvalFunc(pState, progress);
@@ -398,7 +391,28 @@ namespace Tunny.Solver
             return result;
         }
 
-        private void SetTrialResultLog(int trialNum, TrialGrasshopperItems result, OptimizationHandlingInfo optInfo, double[] parameter)
+        private void SetOptimizationParameter(Parameter[] parameter, dynamic trial)
+        {
+            foreach ((VariableBase variable, int i) in Variables.Select((v, i) => (v, i)))
+            {
+                string name = variable.NickName;
+                switch (variable)
+                {
+                    case NumberVariable number:
+                        double numParam = number.IsInteger
+                            ? trial.suggest_int(name, number.LowerBond, number.UpperBond, step: number.Epsilon)
+                            : trial.suggest_float(name, number.LowerBond, number.UpperBond, step: number.Epsilon);
+                        parameter[i] = new Parameter(numParam);
+                        break;
+                    case CategoricalVariable category:
+                        string catParam = (string)trial.suggest_categorical(name, category.Categories);
+                        parameter[i] = new Parameter(catParam);
+                        break;
+                }
+            }
+        }
+
+        private void SetTrialResultLog(int trialNum, TrialGrasshopperItems result, OptimizationHandlingInfo optInfo, Parameter[] parameter)
         {
             var sb = new StringBuilder();
             sb.Append("Trial " + trialNum + " finished with values: {");
@@ -470,7 +484,7 @@ namespace Tunny.Solver
             return isOptimizeCompleted;
         }
 
-        private ProgressState SetProgressState(OptimizationHandlingInfo optSet, double[] parameter, int trialNum, DateTime startTime)
+        private ProgressState SetProgressState(OptimizationHandlingInfo optSet, Parameter[] parameter, int trialNum, DateTime startTime)
         {
             ComputeBestValues(optSet.Study, trialNum, out double[][] bestValues, out double hypervolumeRatio);
             return new ProgressState
@@ -478,7 +492,7 @@ namespace Tunny.Solver
                 TrialNumber = trialNum,
                 ObjectiveNum = Objective.Length,
                 BestValues = bestValues,
-                Values = parameter.Select(v => (decimal)v).ToList(),
+                Parameter = parameter,
                 HypervolumeRatio = hypervolumeRatio,
                 EstimatedTimeRemaining = optSet.Timeout <= 0
                     ? TimeSpan.FromSeconds((DateTime.Now - startTime).TotalSeconds * (optSet.NTrials - trialNum) / (trialNum + 1))