+++ Tunny/Util/GrasshopperInOut.cs
@@ -30,11 +30,12 @@ namespace Tunny.Util
         private readonly GH_Component _component;
         private List<GalapagosGeneListObject> _genePool;
         private GH_FishAttribute _attributes;
-        private List<GH_NumberSlider> Sliders { get; set; }
+        private List<GH_NumberSlider> _sliders;
+        private List<GH_ValueList> _valueLists;
 
         public string ComponentFolder { get; }
         public Objective Objectives { get; private set; }
-        public List<Variable> Variables { get; private set; }
+        public List<VariableBase> Variables { get; private set; }
         public Artifact Artifacts { get; private set; }
         public Dictionary<string, FishEgg> EnqueueItems { get; private set; }
         public bool HasConstraint { get; private set; }
@@ -54,26 +55,33 @@ namespace Tunny.Util
 
         private bool SetVariables()
         {
-            Sliders = new List<GH_NumberSlider>();
+            _sliders = new List<GH_NumberSlider>();
+            _valueLists = new List<GH_ValueList>();
             _genePool = new List<GalapagosGeneListObject>();
 
             _inputGuids.AddRange(_component.Params.Input[0].Sources.Select(source => source.InstanceGuid));
             if (_inputGuids.Count == 0)
             {
-                TunnyMessageBox.Show(
-                    "No input variables found. \nPlease connect a number slider to the input of the component.",
-                    "Tunny",
-                    MessageBoxButtons.OK,
-                    MessageBoxIcon.Error);
+                NoVariableInputError();
                 return false;
             }
 
-            var variables = new List<Variable>();
+            var variables = new List<VariableBase>();
             if (!FilterInputVariables()) { return false; }
             SetInputSliderValues(variables);
             SetInputGenePoolValues(variables);
+            SetInputValueItem(variables);
             Variables = variables;
             return true;
+        }
+
+        private static void NoVariableInputError()
+        {
+            TunnyMessageBox.Show(
+                "No input variables found. \nPlease connect a number slider to the input of the component.",
+                "Tunny",
+                MessageBoxButtons.OK,
+                MessageBoxIcon.Error);
         }
 
         private bool FilterInputVariables()
@@ -84,10 +92,13 @@ namespace Tunny.Util
                 switch (docObject)
                 {
                     case GH_NumberSlider slider:
-                        Sliders.Add(slider);
+                        _sliders.Add(slider);
                         break;
                     case GalapagosGeneListObject genePool:
                         _genePool.Add(genePool);
+                        break;
+                    case GH_ValueList valueList:
+                        _valueLists.Add(valueList);
                         break;
                     case Param_FishEgg fishEgg:
                         if (fishEgg.SourceCount != 0)
@@ -123,11 +134,11 @@ namespace Tunny.Util
             return false;
         }
 
-        private void SetInputSliderValues(List<Variable> variables)
+        private void SetInputSliderValues(List<VariableBase> variables)
         {
             int i = 0;
 
-            foreach (GH_NumberSlider slider in Sliders)
+            foreach (GH_NumberSlider slider in _sliders)
             {
                 decimal min = slider.Slider.Minimum;
                 decimal max = slider.Slider.Maximum;
@@ -167,11 +178,11 @@ namespace Tunny.Util
                         break;
                 }
 
-                variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, nickName, eps, Convert.ToDouble(value), id));
+                variables.Add(new NumberVariable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, nickName, eps, Convert.ToDouble(value), id));
             }
         }
 
-        private void SetInputGenePoolValues(List<Variable> variables)
+        private void SetInputGenePoolValues(List<VariableBase> variables)
         {
             var nickNames = new List<string>();
             for (int i = 0; i < _genePool.Count; i++)
@@ -190,8 +201,20 @@ namespace Tunny.Util
                     IGH_Goo[] goo = genePool.VolatileData.AllData(false).ToArray();
                     var ghNumber = (GH_Number)goo[j];
                     string name = nickNames[i] + j;
-                    variables.Add(new Variable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, name, eps, ghNumber.Value, id));
+                    variables.Add(new NumberVariable(Convert.ToDouble(lowerBond), Convert.ToDouble(upperBond), isInteger, name, eps, ghNumber.Value, id));
                 }
+            }
+        }
+
+        private void SetInputValueItem(List<VariableBase> variables)
+        {
+            foreach (GH_ValueList valueList in _valueLists)
+            {
+                string nickName = valueList.NickName;
+                Guid id = valueList.InstanceGuid;
+                string[] categories = valueList.ListItems.Select(value => value.Name).ToArray();
+                string selectedItem = valueList.FirstSelectedItem.Name;
+                variables.Add(new CategoricalVariable(categories, selectedItem, nickName, id));
             }
         }
 
@@ -284,11 +307,31 @@ namespace Tunny.Util
             return false;
         }
 
-        private bool SetSliderValues(IList<decimal> parameters)
+        private void SetCategoryValues(string[] categoryParameters)
+        {
+            int i = 0;
+            foreach (GH_ValueList valueList in _valueLists)
+            {
+                if (valueList == null)
+                {
+                    return;
+                }
+                string[] categories = valueList.ListItems.Select(item => item.Name).ToArray();
+                int index = Array.IndexOf(categories, categoryParameters[i++]);
+                if (index == -1)
+                {
+                    return;
+                }
+                valueList.SelectItem(index);
+                valueList.ExpireSolution(false);
+            }
+        }
+
+        private bool SetSliderValues(decimal[] parameters)
         {
             int i = 0;
 
-            foreach (GH_NumberSlider slider in Sliders)
+            foreach (GH_NumberSlider slider in _sliders)
             {
                 if (slider == null)
                 {
@@ -342,9 +385,12 @@ namespace Tunny.Util
             while (_document.SolutionState != GH_ProcessStep.PostProcess || _document.SolutionDepth != 0) { }
         }
 
-        public void NewSolution(IList<decimal> parameters)
+        public void NewSolution(IList<Parameter> parameters)
         {
-            SetSliderValues(parameters);
+            decimal[] decimalParameters = parameters.Where(p => p.HasNumber).Select(p => (decimal)p.Number).ToArray();
+            string[] categoryParameters = parameters.Where(p => p.HasCategory).Select(p => p.Category).ToArray();
+            SetSliderValues(decimalParameters);
+            SetCategoryValues(categoryParameters);
             Recalculate();
             SetObjectives();
             SetAttributes();